Backend/books_backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
django-filter==23.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
Pillow==10.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
